Since we are using standalone components,
Now  in main.ts with bootstrapApplication we can add provideHttpClient()

Question: We added provideHttpClient() in app.config.ts not in main.ts ? Is this ok ? What is the difference ?
Difference explained:
main.ts:
You can add providers directly in bootstrapApplication(AppComponent, { providers: [...] }). This is the entry point of your app.

app.config.ts:
This file is used to centralize and organize your application's providers and configuration. You then import and spread these providers in main.ts when bootstrapping your app.